+++ include/iocore/cache/CacheEvacuateDocVC.h
@@ -39,7 +39,7 @@
 #include "tscore/ink_assert.h"
 
 // ts
-#include "ts/DbgCtl.h"
+#include "api/DbgCtl.h"
 
 class CacheEvacuateDocVC : public CacheVC
 {
