+++ include/tscore/Diags.h
@@ -33,9 +33,9 @@
 
 #pragma once
 
-#include <ts/DbgCtl.h>
+#include <api/DbgCtl.h>
+#include "api/SourceLocation.h"
 #include "tscore/DiagsTypes.h"
-#include "tscore/SourceLocation.h"
 #include "tscore/LogMessage.h"
 
 //////////////////////////////////////////////////////////////////////////
