+++ include/tscore/DiagsTypes.h
@@ -40,7 +40,7 @@
 #include "tscore/ink_inet.h"
 #include "tscore/ink_mutex.h"
 #include "tscpp/util/Regex.h"
-#include "tscore/SourceLocation.h"
+#include "api/SourceLocation.h"
 
 #include "tscpp/util/ts_diag_levels.h"
